Proyecto/v2/requirements.txt
Flask>=2.2
Flask-Mail==0.9.1
psycopg2-binary
gunicorn
numpy
pandas
scikit-learn